Fixing casts

+++ indexed_permutation_generator/indexed_permutation_generator.cpp
@@ -11,27 +11,27 @@ unsigned long Factorial(unsigned long n) {
         return 1;
 }
 
-vector<long> Permutation(long element_count, long permutation_index) {
-        vector<long> element_vector, perm;
+vector<unsigned long> Permutation(unsigned long element_count, unsigned long permutation_index) {
+        vector<unsigned long> element_vector, perm;
 
-        long total = Factorial(element_count);
+        unsigned long total = Factorial(element_count);
         permutation_index %= static_cast<unsigned int>(total); // avoiding Invalid permutation. Valid values: 0 <= permutation < factorial(elements)
 
-        for(int i = 0; i < element_count; ++i) {
+        for(unsigned long i = 0; i < element_count; ++i) {
                 element_vector.push_back(i);
         }
 
-        for(int i = 1; i < element_count; ++i) {
+        for(unsigned long i = 1; i < element_count; ++i) {
                 total /= (element_count - i + 1);	// Same as total = Factorial(elements - i), but faster.
                 perm.push_back(element_vector[static_cast<unsigned long>(permutation_index/total)]);
-                element_vector.erase(element_vector.begin() + permutation_index/total);
+                element_vector.erase(element_vector.begin() + static_cast<long>(permutation_index/total));
                 permutation_index %= total;
         }
         perm.push_back(element_vector[0]);
         return perm;
 }
 
-unsigned long ElementPosition(const vector<long> &set, long element) {
+unsigned long ElementPosition(const vector<unsigned long> &set, unsigned long element) {
         for(unsigned long i = 0; i < set.size(); ++i) {
                 if(set[i] == element) {
                         return i;
@@ -40,13 +40,13 @@ unsigned long ElementPosition(const vector<long> &set, long element) {
         return set.size();
 }
 
-unsigned long Index(const vector<long> &permutation) {
+unsigned long Index(const vector<unsigned long> &permutation) {
         // Valid permutatios have:
         //	1: size() = n
         //	2:Non repeated and scrambled elements 0, 1, 2, 3, ..., n-1
-        vector<long> element_vector, perm;
-        for(unsigned int i = 0; i < permutation.size(); ++i) {
-                element_vector.push_back(static_cast<int>(i));
+        vector<unsigned long> element_vector, perm;
+        for(unsigned long i = 0; i < permutation.size(); ++i) {
+                element_vector.push_back(i);
         }
 
         unsigned long index = 0;
@@ -60,7 +60,7 @@ unsigned long Index(const vector<long> &permutation) {
         return index;
 }
 
-void PrintPermutation(const vector<long> &permutation) {
+void PrintPermutation(const vector<unsigned long> &permutation) {
         for(unsigned long i = 0; i <permutation.size() ; ++i) {
                 if(i) {
                         cout << ", ";
@@ -71,7 +71,7 @@ void PrintPermutation(const vector<long> &permutation) {
 
 int main(int, char **) {
         vector<long> permutation;
-        long elements = 9;
+        unsigned long elements = 9;
 
         for(unsigned long i = 0; i < Factorial(elements); ++i) {
                 cout << "Permutation " << i << ": ";
@@ -81,3 +81,4 @@ int main(int, char **) {
 
         return 0;
 }
+
